For Python testing artifacts, introduce combination of Concourse cache and pip --cache-dir.
For the Python testing artifacts used by the CLI tools, utilize the
Concourse cached directories feature to create and use a pip cache dir
shared between task runs.

Be aware, the cache is scoped to the worker the task is run on. We do
not get a cache hit when subsequent builds run on different workers.

* The environment variable PIP_CACHE_DIR is used to store the cache
directory.

* Add "--retries 10" to Behave test dependency pip install commands.