gpstart: improve handling of down segment hosts
Currently, if a host is unreachable when gpstart is run, it fails with an error
regarding inconsistent filespaces, which is both inaccurate and unhelpful to the
user.

This commit adds a check to verify that all hosts are reachable before beginning
the startup proess and, if one or more hosts are not reachable, marks segments
on those hosts down (in gparray, not in the cluster) so gpstart won't try to run
any checks against unreachable hosts.